+++ migrations/20161123125156_subscription_schema.js
@@ -5,7 +5,7 @@ exports.up = (knex, Promise) => {
     table.string('last_name').notNullable();
     table.string('invoice').notNullable();
     table.string('frequency').notNullable();
-    table.bigInteger('amount').notNullable();
+    table.decimal('amount').notNullable();
     table.bigInteger('created_at').notNullable().defaultTo(Date.now());
   });
 };
