+++ migrations/20161123181610_invoice_schema.js
@@ -3,7 +3,7 @@ exports.up = (knex, Promise) => {
     table.increments();
     table.string('first_name').notNullable();
     table.string('last_name').notNullable();
-    table.bigInteger('amount').notNullable();
+    table.decimal('amount').notNullable();
     table.bigInteger('created_at').notNullable().defaultTo(Date.now());
   });
 };
